Move bytecode only startup code out of startup_aux.c and into startup_byt.c
