+++ runtime/caml/startup_aux.h
@@ -61,8 +61,6 @@ struct caml_params {
 extern const struct caml_params* const caml_params;
 
 extern void caml_parse_ocamlrunparam (void);
-extern int caml_parse_command_line (char_os **argv);
-extern void caml_command_error(char *msg, ...);
 
 /* Common entry point to caml_startup.
    Returns 0 if the runtime is already initialized.
