+++ runtime/startup_aux.c
@@ -29,7 +29,6 @@
 #include "caml/osdeps.h"
 #include "caml/startup_aux.h"
 #include "caml/prims.h"
-#include "caml/version.h"
 
 #ifdef _WIN32
 extern void caml_win32_unregister_overflow_detection (void);
@@ -198,112 +197,3 @@ void caml_init_section_table(const char* section_table,
   params.section_table = section_table;
   params.section_table_size = section_table_size;
 }
-
-#ifndef NATIVE_CODE
-
-static void do_print_help(void)
-{
-  printf("%s\n",
-    "Usage: ocamlrun [<options>] [--] <executable> [<command-line>]\n"
-    "Options are:\n"
-    "  -b  Set runtime parameter b (detailed exception backtraces)\n"
-    "  -config  Print configuration values and exit\n"
-    "  -I <dir>  Add <dir> to the list of DLL search directories\n"
-    "  -m  Print the magic number of <executable> and exit\n"
-    "  -M  Print the magic number expected by this runtime and exit\n"
-    "  -p  Print the names of the primitives known to this runtime\n"
-    "  -t  Trace the execution of the bytecode interpreter (specify multiple\n"
-    "      times to increase verbosity)\n"
-    "  -v  Set runtime parameter v=61 (GC event information)\n"
-    "  -version  Print version string and exit\n"
-    "  -vnum  Print short version number and exit\n"
-    "  -help  Display this list of options\n"
-    "  --help  Display this list of options");
-}
-
-/* Print the specified error message followed by an end-of-line and exit */
-extern void caml_command_error(char *msg, ...)
-{
-  va_list ap;
-  va_start(ap, msg);
-  vfprintf (stderr, msg, ap);
-  va_end(ap);
-  fprintf(stderr, "\n");
-  exit(127);
-}
-
-/* Parse options on the command line */
-
-extern int caml_parse_command_line(char_os **argv)
-{
-  int i, j, len, parsed;
-
-  for(i = 1; argv[i] != NULL && argv[i][0] == '-'; i++) {
-    len = strlen_os(argv[i]);
-    parsed = 1;
-    if (len == 2) {
-      /* Single-letter options, e.g. -v */
-      switch(argv[i][1]) {
-      case '-':
-        return i + 1;
-        break;
-      case 't':
-        params.trace_level += 1; /* ignored unless DEBUG mode */
-        break;
-      case 'v':
-        params.verb_gc = 0x001+0x004+0x008+0x010+0x020;
-        break;
-      case 'p':
-        for (j = 0; caml_names_of_builtin_cprim[j] != NULL; j++)
-          printf("%s\n", caml_names_of_builtin_cprim[j]);
-        exit(0);
-        break;
-      case 'b':
-        caml_record_backtraces(1);
-        break;
-      case 'I':
-        if (argv[i + 1] != NULL) {
-          caml_ext_table_add(&caml_shared_libs_path, argv[i + 1]);
-          i++;
-        } else {
-          caml_command_error("option '-I' needs an argument.");
-        }
-        break;
-      case 'm':
-        params.print_magic = 1;
-        break;
-      case 'M':
-        printf("%s\n", EXEC_MAGIC);
-        exit(0);
-        break;
-      default:
-        parsed = 0;
-      }
-    } else {
-      /* Named options, e.g. -version */
-      if (!strcmp_os(argv[i], T("-version"))) {
-        printf("%s\n", "The OCaml runtime, version " OCAML_VERSION_STRING);
-        printf ("The OCaml runtime, version " OCAML_VERSION_STRING "\n");
-        exit(0);
-      } else if (!strcmp_os(argv[i], T("-vnum"))) {
-        printf("%s\n", OCAML_VERSION_STRING);
-        exit(0);
-      } else if (!strcmp_os(argv[i], T("-help")) ||
-                 !strcmp_os(argv[i], T("--help"))) {
-        do_print_help();
-        exit(0);
-      } else if (!strcmp_os(argv[i], T("-config"))) {
-        params.print_config = 1;
-      } else {
-        parsed = 0;
-      }
-    }
-
-    if (!parsed)
-      caml_command_error("unknown option %s", caml_stat_strdup_of_os(argv[i]));
-  }
-
-  return i;
-}
-
-#endif /* not NATIVE_CODE */
